requirements.txt
beautifulsoup4==4.12.3
elasticsearch==8.15.1
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-text-splitters==0.3.0
openai==1.45.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.1.0
streamlit==1.38.0
tiktoken==0.7.0
tqdm==4.66.5